ep 16.  Textbox done
 - text area trigger input
 - box trigger input
 - input change to include parent entities
