--- opengl_build_rpg_tutorial/component_trigger_clean_engine.h
@@ -1,29 +1,0 @@
-#pragma once
-#include "component_trigger.h"
-#include "component_grouped_objects.h"
-
-namespace Component {
-	namespace Trigger
-	{
-		class CleanEngine : public ITrigger
-		{
-		public:
-			void execute(Entity* gamestate) override
-			{
-				auto engine = gamestate->get_child("engine");
-				auto grouped_render_systems = engine->get_component<Component::GroupedSystems>("render");
-				auto grouped_update_systems = engine->get_component<Component::GroupedSystems>("update");
-
-				for (auto grs : grouped_render_systems->objects)
-					for (auto i = 0u; i < grs.size; ++i)
-						if (grs[i]->is_removed())
-							grs.remove(i);
-
-				for (auto gus : grouped_update_systems->objects)
-					for (auto i = 0u; i < gus.size; ++i)
-						if (gus[i]->is_removed())
-							gus.remove(i);
-			}
-		};
-	}
-}
