+++ opengl_build_rpg_tutorial/component_trigger_input_camera.h
@@ -3,8 +3,6 @@
 #include "component_transform.h"
 #include "rect.h"
 #include <glm/vec2.hpp>
-#include "component_observer.h"
-#include "component_trigger_input_observer.h"
 
 /*
 Creates our local camera using width and height
@@ -29,9 +27,6 @@ namespace Component {
 			private:
 				void create(Entity* gamestate) override final
 				{
-					auto& cti_observer = *gamestate->get_component<Component::Trigger::Input::SystemObs>(0);
-					cti_observer.add_subscriber(name_);
-
 					entity_->add_component<Component::Transform>(Rect{ 0.0f, 0.0f, width_, height_ });
 				}
