+++ opengl_build_rpg_tutorial/component_trigger_input_controller.h
@@ -22,9 +22,6 @@ namespace Component {
 			private:
 				void create(Entity* gamestate) override final
 				{
-					auto& cti_observer = *gamestate->get_component<Component::Trigger::Input::SystemObs>(0);
-					cti_observer.add_subscriber(name_);
-
 					entity_->add_component<Component::Controller::Keyboard>();
 				}
 			};
