+++ opengl_build_rpg_tutorial/component_trigger_input_engine.h
@@ -2,8 +2,6 @@
 #include "component_vector.h"
 #include "component_trigger_input.h"
 #include "component_system.h"
-#include "component_grouped_objects.h"
-#include "component_trigger_clean_engine.h"
 
 /*
 Holds our render and update queues for our game engine
@@ -28,8 +26,6 @@ namespace Component {
 				{
 					entity_->add_id_component<Component::SystemVector>("update");
 					entity_->add_id_component<Component::SystemVector>("render");
-					entity_->add_id_component<Component::TriggerVector>("trigger");
-					entity_->add_component<Component::Trigger::CleanEngine>();
 				}
 			};
 		}
