+++ opengl_build_rpg_tutorial/component_trigger_input_font.h
@@ -28,9 +28,6 @@ namespace Component {
 			private:
 				void create(Entity* gamestate) override final
 				{
-					auto& cti_observer = *gamestate->get_component<Component::Trigger::Input::SystemObs>(0);
-					cti_observer.add_subscriber(name_);
-
 					// load tilemap from file
 					std::stringstream fnt_stream;
