+++ opengl_build_rpg_tutorial/component_trigger_input_gamestate_overworld.h
@@ -12,7 +12,7 @@
 #include "component_trigger_input_dependent_collider_map.h"
 #include "component_trigger_input_font.h"
 #include "component_trigger_input_dependent_text.h"
-#include "component_trigger_input_dependent_sandbox.h"
+#include "component_trigger_input_dependent_textbox.h"
 
 /*
 Set up class for all game object creation within the overworld state
@@ -35,7 +35,6 @@ namespace Component {
 				private:
 					void create(Entity* gamestate) override final
 					{
-						entity_->push_back_component<Component::Trigger::Input::SystemObs>("observer");
 						entity_->push_back_component<Component::Trigger::Input::Engine>("engine", 64u);
 						entity_->push_back_component<Component::Trigger::Input::Camera>("camera", 64.0f * 32.0f);
 						entity_->push_back_component<Component::Trigger::Input::Controller>("controller");
@@ -47,8 +46,7 @@ namespace Component {
 						entity_->push_back_component<Component::Trigger::Input::Dependent::TileMap>("tilemap", "resources/data/TestTileMapGJK.json");
 						entity_->push_back_component<Component::Trigger::Input::Dependent::Player>("player", (GLfloat)Game::width, 792.0f);
 						entity_->push_back_component<Component::Trigger::Input::Dependent::ColliderMap>("collider map");
-						entity_->push_back_component<Component::Trigger::Input::Dependent::Sandbox>("sandbox");
-						//entity_->push_back_component<Component::Trigger::Input::Dependent::Text>("display", "Hello world!", glm::vec2(30.0f, 20.0f));
+						entity_->push_back_component<Component::Trigger::Input::Dependent::TextBox>("textbox");
 					}
 				};
 			}
