+++ opengl_build_rpg_tutorial/component_trigger_input_renderer.h
@@ -26,9 +26,6 @@ namespace Component {
 			private:
 				void create(Entity* gamestate) override final
 				{
-					auto& cti_observer = *gamestate->get_component<Component::Trigger::Input::SystemObs>(0);
-					cti_observer.add_subscriber(name_);
-
 					// create a renderer object and input appropriate attribute sizes and max amount of sprites on screen at once
 					// 2 = pos, 2 = coords
 					entity_->add_component<Component::Renderer>(std::vector<GLuint>{ 2, 2 }, MAX_SPRITES);
