+++ opengl_build_rpg_tutorial/opengl_build_rpg_tutorial.vcxproj
@@ -156,9 +156,6 @@
   <ItemGroup>
     <ClInclude Include="component.h" />
     <ClInclude Include="component_array.h" />
-    <ClInclude Include="component_grouped_objects.h" />
-    <ClInclude Include="component_observer.h" />
-    <ClInclude Include="component_doubly_linked_list.h" />
     <ClInclude Include="component_src_bitmap_glyph.h" />
     <ClInclude Include="component_collider.h" />
     <ClInclude Include="component_collider_aabb.h" />
@@ -171,20 +168,11 @@
     <ClInclude Include="component_system_render_draw.h" />
     <ClInclude Include="component_system_update_traverse_tree.h" />
     <ClInclude Include="component_template.h" />
-    <ClInclude Include="component_tree.h" />
-    <ClInclude Include="component_trigger_add_game_obj.h" />
-    <ClInclude Include="component_trigger_clean_engine.h" />
-    <ClInclude Include="component_trigger_delete_entity.h" />
-    <ClInclude Include="component_trigger_delete_game_obj.h" />
-    <ClInclude Include="component_trigger_input_dependent_sandbox.h" />
     <ClInclude Include="component_trigger_input_dependent_box.h" />
     <ClInclude Include="component_trigger_input_dependent_text_area.h" />
     <ClInclude Include="component_trigger_input_dependent_text.h" />
     <ClInclude Include="component_trigger_input_font.h" />
     <ClInclude Include="component_trigger_input_dependent_textbox.h" />
-    <ClInclude Include="component_trigger_input_observer.h" />
-    <ClInclude Include="component_trigger_switch_systems.h" />
-    <ClInclude Include="doubly_linked_list.h" />
     <ClInclude Include="frarr.h" />
     <ClInclude Include="physics.h" />
     <ClInclude Include="component_collider_gjk_physics_boundary_smooth.h" />
